pyproject.toml
[tool.poetry]
name = "luthor-backend"
version = "0.1.0"
description = ""
authors = ["wmartzh <wmartzh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
